% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{AddAdj}
\alias{AddAdj}
\title{Add adjacency matrix list for a SRTProject object}
\usage{
AddAdj(
  SRTProj,
  platform = c("Visium", "ST", "Other"),
  type = "fixed_distance",
  force = FALSE,
  ...
)
}
\arguments{
\item{SRTProj}{a SRTProject object}

\item{platform}{a string, specify the data collection platform, one of "Visium", "ST" and "Other", default as "Visium". The platform helps to calculate the adjacency matrix by defining the neighborhoods when type="fixed_distance" is chosen.}

\item{type}{an optional string, specify which type of neighbors' definition. Here we provide two definition: one is "fixed_distance", the other is "fixed_number".}

\item{force}{a logical value, whether delete and rewrite the adjacency matrix list to h5file if there exists the adjacency matrix.}

\item{...}{other arguments pass to \code{\link{getAdj_auto}}.}
}
\value{
return a SRTProject.
}
\description{
Add adjacency matrix list for a SRTProject object
}
