% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{AddClusters2MetaData}
\alias{AddClusters2MetaData}
\title{Add inferred clusters to the \code{cellMetaData}}
\usage{
AddClusters2MetaData(SRTProj, name = "Clusters")
}
\arguments{
\item{SRTProj}{a SRTProject object after fitting clustering method.}

\item{name}{a string, the name created in colData to put the cluster labels.}
}
\value{
return a revised SRTProject.
}
\description{
Add inferred clusters to the \code{cellMetaData}
}
