% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{AddPCA}
\alias{AddPCA}
\alias{AddPCA.SRTProject}
\title{Add PCA embeddings for a SRTProject object}
\usage{
AddPCA(obj, n_comp = 15, Method = c("APCA", "PCA", "WPCA"), seed = 1)

\method{AddPCA}{SRTProject}(obj, n_comp = 15, Method = c("APCA", "PCA", "WPCA"), seed = 1)
}
\arguments{
\item{obj}{a \code{SRTProject} object.}

\item{n_comp}{an optional positive integer, specify the number of features to be extracted.}

\item{Method}{a string, the method to compute PCA scores. There are three versions to provide: approximated PCA (APCA), PCA, and weighted PCA (WPCA).}

\item{seed}{a non-negative integer, the random seed for reproducibility.}
}
\value{
Return a revised SRTProject object by adding PCA embeddings
}
\description{
Add PCA embeddings for a SRTProject object
}
