% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{addProjectSummary}
\alias{addProjectSummary}
\title{Add the summary data for a SRTProject object}
\usage{
addProjectSummary(SRTProj = NULL, name = NULL, summary = NULL)
}
\arguments{
\item{SRTProj}{a SRTProject object.}

\item{name}{a string, set a name for the new summary data to be added.}

\item{summary}{any object, the new summary data to be added.}
}
\value{
return a revised SRTProject object.
}
\description{
Add the summary data for a SRTProject object
}
