% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{AddTrajectory}
\alias{AddTrajectory}
\title{Implement trajectory inference}
\usage{
AddTrajectory(
  spe,
  use_cluster = "clusters",
  reduction = NULL,
  name = NULL,
  traject.method = "Slingshot",
  target.clus = NULL,
  start.clus = NULL,
  seed = 1,
  return.all.branch = FALSE,
  ...
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{use_cluster}{a string, the name of clusters in the colData of the  SpatialExperiment object used for trajectory inference.}

\item{reduction}{a string, the name of reduction slot used in trajectory inference.}

\item{name}{a string, the name of inferrend pseudotime, default as "PT".}

\item{traject.method}{a string, the method used to conduct trajectory inference, default as 'Slingshot'.}

\item{target.clus}{a string, the cluster to be used as end point.}

\item{start.clus}{a string, the cluster to be used as the start point.}

\item{seed}{a positive integer, the random seed.}

\item{return.all.branch}{a logical value, whether return all inferred braches, default as \code{FALSE} implies only the merged pseudotime is returned.}

\item{...}{other arguments pass to other method.}
}
\value{
return a revised SpatialExperiment object. The colData of the SpatialExperiment object adds a column named name (PT).
}
\description{
Implement trajectory inference based on low-dimensional embeddings and clusters using Slingshot or other methods.
}
