% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{AddTSNE}
\alias{AddTSNE}
\alias{AddTSNE.SRTProject}
\title{Add tSNE embeddings for a SRTProject object}
\usage{
AddTSNE(SRTProj, n_comp = 2, reduction = NULL, seed = 1)

\method{AddTSNE}{SRTProject}(SRTProj, n_comp = 2, reduction = NULL, seed = 1)
}
\arguments{
\item{SRTProj}{a \code{SRTProject} object.}

\item{n_comp}{an optional positive integer, specify the number of features to be extracted.}

\item{reduction}{an optional string, means which dimensional reduction (e.g. PRECAST, PCA) to use for the tSNE, default as the last one in the SRTProject object.}

\item{seed}{a non-negative integer, the random seed for reproducibility.}
}
\value{
Return a revised SRTProject object by adding tSNE embeddings
metadata
}
\description{
Add tSNE embeddings for a SRTProject object
}
