% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{AddUMAP}
\alias{AddUMAP}
\alias{AddUMAP.SRTProject}
\title{Add UMAP embeddings for a SRTProject object}
\usage{
AddUMAP(SRTProj, n_comp = 2, reduction = NULL, seed = 1)

\method{AddUMAP}{SRTProject}(SRTProj, n_comp = 2, reduction = NULL, seed = 1)
}
\arguments{
\item{SRTProj}{a \code{SRTProject} object.}

\item{n_comp}{an optional positive integer, specify the number of features to be extracted.}

\item{reduction}{an optional string, means which dimensional reduction (e.g. PRECAST, PCA) to use for the UMAP, default as the last one in the SRTProject object.}

\item{seed}{a non-negative integer, the random seed for reproducibility.}
}
\value{
Return a revised SRTProject object by adding UMAP embeddings
}
\description{
Add UMAP embeddings for a SRTProject object
}
