% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_newinterface.R
\name{Annotation_SpatialAnno}
\alias{Annotation_SpatialAnno}
\title{Make annotations for SRT data by fitting a SpatialAnno model}
\usage{
Annotation_SpatialAnno(SRTProj, markers, q = 15, ...)
}
\arguments{
\item{SRTProj}{an object named "SRTProject". The object SRTProject is created by \link{CreateSRTProject}.}

\item{markers}{a named string vector, the marker genes used for annotation.}

\item{q}{An optional integer, specify the number of low-dimensional embeddings to extract in DR-SC model.}

\item{...}{other parameters pass to \link{model_set_SpatialAnno}.
\href{https://www.biorxiv.org/content/10.1101/2023.02.08.527590v1}{Shi, X., Yang, Y., Ma, X., Zhou, Y., Guo, Z., Wang, C., & Liu, J. (2023). Probabilistic cell/domain-type assignment of spatial transcriptomics data with SpatialAnno. bioRxiv, 2023-02.}}
}
\description{
Make annotations for SRT data by fitting a SpatialAnno model
}
