% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{barPlot_enrich}
\alias{barPlot_enrich}
\title{Barplot for the enrichment terms}
\usage{
barPlot_enrich(
  top_dat,
  source = "Ont",
  term_name = "Term",
  nlog10P = "nlog10P",
  bar_width = 0.8,
  base_size = 20,
  font_family = "serif",
  cols = (ggthemes::canva_pal())(4)
)
}
\arguments{
\item{top_dat}{a data frame with the information of enrichment terms}

\item{source}{a string, the source of database.}

\item{term_name}{a string, the name of y-axis.}

\item{nlog10P}{a string, the colname in top_dat that represents the negative log10 p-values.}

\item{bar_width}{a positive real, the width of bars.}

\item{base_size}{a positive real, the  baseline text size in plot.}

\item{font_family}{a string, the font.}

\item{cols}{a string vector, the used colors in plot.}
}
\value{
return a ggplot object.
}
\description{
Barplot for the enrichment terms
}
