% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CCHeatMap}
\alias{CCHeatMap}
\alias{CCHeatMap.SRTProject}
\title{Plot cluster-by-cluster correlation hetamap}
\usage{
CCHeatMap(
  obj,
  reduction = NULL,
  legend_title = "Domain",
  grp_label = FALSE,
  pt_size = 4,
  grp_color = NULL,
  fill_legend_title = "Pearson's correlation",
  seed = 1,
  ...
)

\method{CCHeatMap}{SRTProject}(
  obj,
  reduction = NULL,
  legend_title = "Domain",
  grp_label = FALSE,
  pt_size = 4,
  grp_color = NULL,
  fill_legend_title = "Pearson's correlation",
  seed = 1,
  ...
)
}
\arguments{
\item{obj}{A SRTProject object with low dimension embeddings.}

\item{reduction}{a string, the name of reduction (low-dimensional embeddings).}

\item{legend_title}{a string, the title of legend.}

\item{grp_label}{a logical value, whether display the cluster labels in plot.}

\item{pt_size}{a positive real, the point size in plot.}

\item{grp_color}{a string vector, the colors for each cluster in plot.}

\item{fill_legend_title}{a string, the legend title for filled colors.}

\item{seed}{a positive integer, a ranodm seed.}

\item{...}{other arguments pass to \code{DoHeatmap}.}
}
\value{
Returns a ggplot object.
}
\description{
Visualize the cluster-by-cluster correlation based on a specified low-dimensional embeddings.
}
\seealso{
None.
}
