% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_newinterface.R
\name{Cluster_DRSC}
\alias{Cluster_DRSC}
\title{Joint embedding and clustering for SRT data by fitting an iSC.MEB model a DR-SC model}
\usage{
Cluster_DRSC(SRTProj, K = NULL, q = 15, ...)
}
\arguments{
\item{SRTProj}{an object named "SRTProject". The object SRTProject is created by \link{CreateSRTProject}.}

\item{K}{An optional integer or integer vector, specify the candidates of number of clusters. if K=NULL, it will be set to 4~12.}

\item{q}{An optional integer, specify the number of low-dimensional embeddings to extract in DR-SC model.}

\item{...}{other parameters pass to \link{model_set_DRSC}.}
}
\description{
Joint embedding and clustering for SRT data by fitting an iSC.MEB model a DR-SC model
}
\references{
\href{https://academic.oup.com/nar/article/50/12/e72/6555431}{Wei Liu, Xu Liao, Yi Yang, Huazhen Lin, Joe Yeong, Xiang Zhou, Xingjie Shi and Jin Liu* (2022). Joint dimension reduction and clustering analysis of single-cell RNA-seq and spatial transcriptomics data, Nucleic Acids Research, gkac219}
}
