% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_newinterface.R
\name{Cluster_Louvain}
\alias{Cluster_Louvain}
\title{Louvain clustering for a SRT data}
\usage{
Cluster_Louvain(SRTProj, resolution = 0.8, reduction = "PCA", ...)
}
\arguments{
\item{SRTProj}{an object named "SRTProject". The object SRTProject is created by \link{CreateSRTProject}.}

\item{resolution}{an positive real, specify the resolution.}

\item{reduction}{a string, specify the reduction slot name in SRTProj for fitting Louvain model.}

\item{...}{other parameters pass to \link{model_set_Louvain}.}
}
\description{
Louvain clustering for a SRT data
}
