% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cluster_metric}
\alias{cluster_metric}
\title{Compute adjusted Rand index and normalized mutual information}
\usage{
cluster_metric(hy, y, type = c("ARI", "NMI"))
}
\arguments{
\item{hy}{a vector, the predicted clusters}

\item{y}{a vector, the true clusters}

\item{type}{a string, the type of cluster metric, one of "ARI" and "NMI", default as "ARI".}
}
\description{
Compute adjusted Rand index and normalized mutual information
}
\examples{
y <- rep(1:4, each=100)
hy <- rep(1:5, each=80)
cluster_metric(y, hy)


}
