% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_newinterface.R
\name{Cluster_SCMEB}
\alias{Cluster_SCMEB}
\title{Spatial clustering for SRT data by fitting a SC-MEB model}
\usage{
Cluster_SCMEB(SRTProj, K = NULL, reduction = "PCA", ...)
}
\arguments{
\item{SRTProj}{an object named "SRTProject". The object SRTProject is created by \link{CreateSRTProject}.}

\item{K}{an optional integer or integer vector, specify the candidates of number of clusters. if K=NULL, it will be set to 4~12.}

\item{reduction}{a string, specify the reduction slot name in SRTProj for fitting SC-MEB model.}

\item{...}{other parameters pass to \link{model_set_iSCMEB}.}
}
\description{
Spatial clustering for SRT data by fitting a SC-MEB model
}
\references{
\href{https://academic.oup.com/bib/article/23/1/bbab466/6440124}{Yi Yang, Xingjie Shi, Wei Liu, et.al and Jin Liu* (2021). SC-MEB: spatial clustering with hidden Markov random field using empirical Bayes. Briefings in Bioinformatics}
}
