% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{CreateSRTProject}
\alias{CreateSRTProject}
\title{Create a SRTProject object}
\usage{
CreateSRTProject(
  cntList,
  coordList,
  projectName = "SRTProject",
  cellMetaDataList = NULL,
  sampleMetadata = NULL,
  min.spots = 0,
  min.genes = 0,
  filter.plot = FALSE,
  prefix = "****",
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{cntList}{a list with each component as a count matrix, has names for all components.}

\item{coordList}{a list with spatial coordinates matrix has the same names as that of cntList.}

\item{projectName}{a string, the name of this project, default as "SRTProject".}

\item{cellMetaDataList}{a list or NULL, provide the cell meta data for each data batch.}

\item{sampleMetadata}{a dataframe object with the number of rows same as the length of cntList, provide the sample meta data.}

\item{min.spots}{a non-negative integer, the least number of nonzero spots for a retained gene used to filter spots in the QC.}

\item{min.genes}{a non-negative integer, the least number of nonzero genes for a retained spot used to filter spots in the QC.}

\item{filter.plot}{a logical value, whether plot the QC plots in filtering step.}

\item{prefix}{a string, specify the prefix for the output information.}

\item{force}{a logical value, whether delete then rewrite the data to h5 file if the project exists.}

\item{verbose}{a logical value, whether output the information when creating a SRTProject object.}
}
\description{
Create a SRTProject object based on at least count matrix list and spatial coordinates list.
}
