% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{EachClusterSpaHeatMap}
\alias{EachClusterSpaHeatMap}
\alias{EachClusterSpaHeatMap.list}
\alias{EachClusterSpaHeatMap.SRTProject}
\title{Plot spatial cluster heatmap for each data batch}
\usage{
EachClusterSpaHeatMap(obj, ...)

\method{EachClusterSpaHeatMap}{list}(
  obj,
  cols = NULL,
  title_name = NULL,
  combine = TRUE,
  layout.dim = c(round(length(obj[[2]])/2), 2),
  no_axis = FALSE,
  pt_size = 2,
  point_alpha = 1,
  nrow.legend = NULL,
  base_size = 10,
  border_col = "gray",
  axis_names = c("Coord x", "Coord y"),
  no_guides = FALSE,
  no_axis_name = FALSE,
  legend.position = "bottom",
  common.legend = TRUE
)

\method{EachClusterSpaHeatMap}{SRTProject}(
  obj,
  batch = NULL,
  cols = NULL,
  layout.dim = NULL,
  combine = TRUE,
  no_axis = TRUE,
  ...
)
}
\arguments{
\item{obj}{A SRTProject object.}

\item{...}{Arguments passed to other methods.}

\item{cols}{a string vector, colors used in the plot.}

\item{title_name}{a string, the title of plot.}

\item{combine}{a logical value,  whether plot all on a figure. If TRUE, all figures are plotted; otherwise, return a list with each plot as component. TRUE by default.}

\item{layout.dim}{a 2 vector,  the dimension in the layout of plots when \code{combine = TRUE}.}

\item{no_axis}{a logical value, whether displys the axis.}

\item{batch}{a string, number or NULL, specify the batch name to plot, default as \code{NULL} implying all batches plotted.}
}
\description{
The function \code{EachClusterSpaHeatMap} is used to visualize  clusters on spatial coordinates with the RGB colors for each data batch.
}
