% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{EachEmbedPlot}
\alias{EachEmbedPlot}
\alias{EachEmbedPlot.SpatialExperiment}
\title{Plot the embeddings for each SRT data}
\usage{
EachEmbedPlot(obj, reduction = NULL, batch = NULL, ...)

\method{EachEmbedPlot}{SpatialExperiment}(
  obj,
  reduction = NULL,
  colour_by = NULL,
  batch = NULL,
  combine = TRUE,
  base_size = 15,
  layout.dim = NULL,
  common.legend = TRUE,
  legend.position = "right",
  add_border_box = FALSE,
  no_axis = TRUE,
  ...
)
}
\arguments{
\item{obj}{an object with class \code{SpatialExperiment }.}

\item{reduction}{a string, speficify which reduction slot is used to plot. If \code{reduction=NULL}, the last one is used.}

\item{batch}{a string vector, denoting the names of batches to be plotted, or a integer vector denoting the ID numbers of batches to be plotted.}

\item{...}{other arguments that pass to other function.}

\item{colour_by}{a string, specify the colname in cell metadata to be mapped to colors.}

\item{combine}{a logical value, whether put all plots in one figure.}

\item{base_size}{a positive real, the baseline text size.}

\item{layout.dim}{a 2-dimension integer vector, specify the layout of figure.}

\item{common.legend}{a logical value, whether use a common legend for all plots.}

\item{legend.position}{a string, the position of legend.}

\item{add_border_box}{a logical value, whether add the border box.}

\item{no_axis}{a logical value, whether display the axis.}
}
\value{
return a ggplot object of a list of ggplot objects.
}
\description{
Plot the embeddings for each SRT data
}
