% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{EachExprSpaHeatMap}
\alias{EachExprSpaHeatMap}
\alias{EachExprSpaHeatMap.list}
\alias{EachExprSpaHeatMap.SRTProject}
\title{Plot spatial feature heatmaps for a group of SRT data batches}
\usage{
EachExprSpaHeatMap(
  obj,
  features,
  batch = NULL,
  quantVec = NULL,
  combine = TRUE,
  scale = TRUE,
  ...
)

\method{EachExprSpaHeatMap}{list}(
  obj,
  quantVec = NULL,
  title_name = TRUE,
  combine = TRUE,
  layout.dim = c(round(length(obj[[2]])), ncol(obj[[2]][[1]])),
  no_axis = TRUE,
  pt_size = 1,
  base_size = 10,
  border_col = "gray",
  common.legend = TRUE,
  axis_names = c("Coord x", "Coord y"),
  no_axis_name = TRUE,
  legend.position = "right",
  ...
)

\method{EachExprSpaHeatMap}{SRTProject}(
  obj,
  features,
  batch = NULL,
  quantVec = NULL,
  combine = TRUE,
  scale = TRUE,
  layout.dim = NULL,
  no_axis = TRUE,
  pt_size = 1,
  base_size = 10,
  border_col = "gray",
  common.legend = TRUE,
  axis_names = c("Coord x", "Coord y"),
  no_axis_name = TRUE,
  legend.position = "right",
  ...
)
}
\arguments{
\item{obj}{a SRTProject object.}

\item{features}{a string vector, a group of genes to be visualized.}

\item{batch}{a string or positive integer vector, the names or ID of batches to be plotted}

\item{quantVec}{a real or real vector between 0 and 1, set the midlle scale of continous color schema.}

\item{combine}{a logical value, whehter put all plots on one figure, default as TRUE.}

\item{scale}{a logical value, whether scale the gene expression for plot, default as \code{TRUE}.}

\item{...}{other arguments pass to \code{\link{ggarrange}}.}

\item{layout.dim}{a 2-diemensional vector, specify the layout (numbers of rows and columns) of figure.}

\item{no_axis}{a logical value, whether display the axis in plot.}

\item{pt_size}{a positive real, the point size in plot.}

\item{base_size}{a positive real, the baseline text size in plot.}

\item{border_col}{a string, the color of border in plot.}

\item{common.legend}{a logcial value, whether use a common legend in figure.}

\item{axis_names}{a 2-dimensional string vector, the axis names in plot.}

\item{no_axis_name}{a logical value, whether display the axis name in plot.}

\item{legend.position}{a string, the position of legend.}
}
\value{
Returns a ggplot object or a list of ggplot objects.
}
\description{
Visualize the gene expressions on spatial coordinates for a group of genes and SRT data batches.
}
\seealso{
None.
}
