% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{EachRGBSpaHeatMap}
\alias{EachRGBSpaHeatMap}
\alias{EachRGBSpaHeatMap.list}
\alias{EachRGBSpaHeatMap.SRTProject}
\title{Plot spatial RGB heatmaps for a group of SRT data batches}
\usage{
EachRGBSpaHeatMap(obj, ...)

\method{EachRGBSpaHeatMap}{list}(
  obj,
  title_name = NULL,
  combine = TRUE,
  layout.dim = c(round(length(obj[[2]])/2), 2),
  no_axis = FALSE,
  pt_size = 2,
  base_size = 10,
  border_col = "gray",
  axis_names = c("Coord x", "Coord y"),
  no_axis_name = FALSE,
  ...
)

\method{EachRGBSpaHeatMap}{SRTProject}(
  obj,
  batch = NULL,
  plot_type = c("UMAP", "tSNE"),
  layout.dim = NULL,
  combine = TRUE,
  title_name = NULL,
  no_axis = TRUE,
  pt_size = 2,
  base_size = 10,
  border_col = "gray",
  axis_names = c("Coord x", "Coord y"),
  no_axis_name = FALSE,
  ...
)
}
\arguments{
\item{obj}{A SRTProject object.}

\item{...}{Arguments passed to \code{\link{ggarrange}}.}

\item{title_name}{a string, the title of plot.}

\item{combine}{a logical value, whether plot all on a figure. If TRUE, all figures are plotted; otherwise, return a list with each plot as component. TRUE by default.}

\item{layout.dim}{a 2 integer vector, the dimension in the layout of plots when \code{combine = TRUE}.}

\item{no_axis}{a logical value, whether display the axis in plot.}

\item{pt_size}{a positive real, the point size in plot.}

\item{base_size}{a positive real, the baseline text size in the plot}

\item{border_col}{a string, the color of border in plot.}

\item{axis_names}{a 2 string vector, the axis names in plot.}

\item{no_axis_name}{a logical value, whether display the axis name in plot.}

\item{batch}{a string vector or positive integer vector, the batch names or IDs to be plotted.}

\item{plot_type}{a string,  the type of plot, one of "UMAP" and "tSNE".}

\item{common.legend}{a logical value,  whether combine the legend of multiple plots. TRUE by default.}
}
\description{
The function \code{EachRGBSpaHeatMap} is used to visualize the RGB colors of 3-dimensional embeddings on spatial coordinates.
}
