% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{EmbedPlot}
\alias{EmbedPlot}
\alias{EmbedPlot.SRTProject}
\alias{EmbedPlot.SpatialExperiment}
\title{Two-dimensional embeddings plot}
\usage{
EmbedPlot(obj, item = "cluster", plotEmbeddings = "tSNE", ...)

\method{EmbedPlot}{SRTProject}(
  obj,
  item = "cluster",
  plotEmbeddings = NULL,
  axis_names = NULL,
  cols = NULL,
  no_guides = FALSE,
  no_axis_name = FALSE,
  pt_size = 1,
  point_alpha = 1,
  nrow.legend = NULL,
  base_size = 10,
  border_col = "gray",
  legend.position = "bottom",
  no_axis = FALSE,
  do_density = FALSE
)

\method{EmbedPlot}{SpatialExperiment}(
  spe,
  plotEmbeddings = NULL,
  colour_by = NULL,
  base_size = 15,
  legend.position = "right",
  add_border_box = FALSE,
  ...
)
}
\arguments{
\item{obj}{a SRTProject object or SpatialExperiment object.}

\item{item}{a string, the item used to mapped as colors.}

\item{plotEmbeddings}{a string, the name of plotembeddings, sucha as tSNE and UMAP.}

\item{axis_names}{a 2-dimensional string vector, the names of axis.}

\item{cols}{a string vector, the selected colors.}

\item{no_guides}{a logical value, whether add guides.}

\item{no_axis_name}{a logical value, whether display the axis names.}

\item{pt_size}{a positive real, the point size in plot.}

\item{point_alpha}{a positive real between 0 and 1, the point transparency in plot.}

\item{nrow.legend}{a positive number, the number of columns for legend.}

\item{base_size}{a positive real, the baseline text size in plot.}

\item{border_col}{a string, the color of border in plot.}

\item{legend.position}{a string, the positive of legend.}

\item{no_axis}{a logical value, whether display the axis.}

\item{do_density}{a logical value, whether plot the density plot.}

\item{add_border_box}{a logical value, whether add border box in plot.}
}
\value{
returns a ggplot object.
}
\description{
Visualize the specified item on the two-dimensional embeddings such as tSNE, UMAP and so on.
}
\seealso{
\code{\link{SRTProject-class}}.
}
