% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_gene}
\alias{filter_gene}
\title{Filter genes for a Seurat object}
\usage{
filter_gene(seu, min_spots = 20, assay = NULL)
}
\arguments{
\item{seu}{a Seurat object.}

\item{min_spots}{a non-negative integer, the least number of nonzero spots for a retained gene.}

\item{assay}{a string, the assay to perform filtering.}
}
\value{
Return a revised Seurat object.
}
\description{
Filter genes for a single-cell RNA sequencing or sptially-resolved transcriptomics data in the form of a Seurat object.
}
