% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_spot}
\alias{filter_spot}
\title{Filter spots for a Seurat object}
\usage{
filter_spot(seu, min_feature = 0, assay = NULL)
}
\arguments{
\item{seu}{a Seurat object.}

\item{min_feature}{a non-negative integer, the least number of nonzero genes for a retained spot.}

\item{assay}{a string, the assay to perform filtering.}
}
\value{
Return a revised Seurat object.
}
\description{
Filter spots for a single-cell RNA sequencing or sptially-resolved transcriptomics data in the form of a Seurat object.
}
