% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{FindAllDEGs}
\alias{FindAllDEGs}
\title{Find differential expression genes for all clusters by comparing one cluster to the remaining clusters.}
\usage{
FindAllDEGs(
  spe,
  use_cluster = "clusters",
  features = NULL,
  logfc.threshold = 0.25,
  only.pos = FALSE,
  test.use = "wilcox",
  ...
)
}
\arguments{
\item{spe}{A SpatialExperiment object with cluster labels for each cell/spot in colData.}

\item{use_cluster}{a string, the name in colData denoting the cluster labels.}

\item{features}{a string vector, specify the gene list to do differential gene expression analysis.}

\item{logfc.threshold}{a real, the threshold of log fold change to determine which genes are returned.}

\item{only.pos}{a logical value, whether only return the genes with positive log fold change.}

\item{test.use}{a string, the test to use.}

\item{...}{other arguments pass to \code{\link{FindAllMarkers}}.}
}
\value{
return a dataframe with the DEG analysis results.
}
\description{

}
