% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{FindDEGs}
\alias{FindDEGs}
\title{Find differential expression genes
Find differential expression genes by comparing two clusters.}
\usage{
FindDEGs(
  spe,
  use_cluster = "clusters",
  cluster1 = NULL,
  cluster2 = NULL,
  features = NULL,
  logfc.threshold = 0.25,
  test.use = "wilcox",
  ...
)
}
\arguments{
\item{spe}{spe A SpatialExperiment object with cluster labels for each cell/spot in colData.}

\item{use_cluster}{a string, the name in colData denoting the cluster labels.}

\item{cluster1}{An optional positive integer, specify the number of features to be extracted.}

\item{cluster2}{A random seed to be used.}

\item{features}{a string vector, specify the gene list to do differential gene expression analysis.}

\item{logfc.threshold}{a real, the threshold of log fold change to determine which genes are returned.}

\item{test.use}{a string, the test to use.}

\item{...}{other arguments pass to \code{\link{FindAllMarkers}}.}
}
\value{
return a dataframe with the DEG analysis results.
}
\description{

}
\seealso{
None
}
