% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{GCHeatMap}
\alias{GCHeatMap}
\alias{GCHeatMap.SpatialExperiment}
\title{Plot gene-by-cell heatmap}
\usage{
GCHeatMap(
  obj,
  features,
  cell_label = "Domain",
  fill_legend_title = "Expression",
  do.scale = TRUE,
  grp_label = FALSE,
  pt_size = 4,
  grp_color = NULL,
  ncol.legend = 1,
  legend.position = "right",
  axis_text_y = TRUE,
  y_text_size = 5,
  ...
)

\method{GCHeatMap}{SpatialExperiment}(
  obj,
  features,
  cell_label = "Domain",
  fill_legend_title = "Expression",
  do.scale = TRUE,
  grp_label = FALSE,
  pt_size = 4,
  grp_color = NULL,
  ncol.legend = 1,
  legend.position = "right",
  axis_text_y = TRUE,
  y_text_size = 5,
  ...
)
}
\arguments{
\item{obj}{a SpatialExperiment object.}

\item{features}{a string vector, a group of selected genes' names.}

\item{cell_label}{a string, specify name of clusters.}

\item{fill_legend_title}{a string, the legend title.}

\item{do.scale}{a logical value, whether scale the gene expression for plot, default as \code{TRUE}.}

\item{grp_label}{a logical value, whether display the cluster labels in plot.}

\item{pt_size}{a positive real, the point size in plot.}

\item{grp_color}{a string vector, the colors for each cluster in plot.}

\item{ncol.legend}{a positive number, the number of columns for legend.}

\item{legend.position}{a string, the postion of legend.}

\item{axis_text_y}{a logical value, whether display the tick text in y-axis.}

\item{y_text_size}{a positive real, the text size for y-axis.}

\item{...}{other arguments pass to \code{DoHeatmap}.}
}
\value{
Returns a ggplot object.

Returns a ggplot object.
}
\description{
Visualize the expressions of the selected features (genes) for all cells/spots.
}
\seealso{
None
}
