% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{get_top_pathway}
\alias{get_top_pathway}
\title{Get top pathways from enrichment analysis results}
\usage{
get_top_pathway(
  df,
  ntop = 10,
  source_set = c("GO:MF", "GO:CC", "GO:BP", "KEGG", "HPA")
)
}
\arguments{
\item{df}{a data frame}

\item{ntop}{a positive integer, how many top terms for each source are returned}

\item{source_set}{a string vector, specify the sources.}
}
\value{
return a dataframe object including the top pathways.
}
\description{
Get top pathways from enrichment analysis results
}
