% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{getGeneSpotData}
\alias{getGeneSpotData}
\title{Get the expression matrix from a SRTProject object}
\usage{
getGeneSpotData(SRTProj)
}
\arguments{
\item{SRTProj}{a SRTProject object.}
}
\value{
return a a SpatialExperiment object with  gene expression matrix and other information from the SRTProject object.
}
\description{
Get the expression matrix from a SRTProject object
}
