% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{getIntegratedData}
\alias{getIntegratedData}
\title{Get the batch-corrected integrated gene expresssions}
\usage{
getIntegratedData(
  SRTProj,
  species = "Human",
  Method = c("iSC.MEB", "PRECAST"),
  custom_housekeep = NULL,
  only.var.features = TRUE,
  add_empty_counts = FALSE
)
}
\arguments{
\item{SRTProj}{a SRTProject object after finishing integration analysis using \code{\link{Integrate_PRECAST}} or \code{\link{Integrate_iSCMEB}}.}

\item{species}{a string, one of "Human", "Mouse" and "Unknown", default as "Human".}

\item{Method}{a string, one of "iSC.MEB" and "PRECAST", specify which method is used to remove the batch effects in the gene expression levels.}

\item{custom_housekeep}{a string vector, a group of user-specified housekeeping genes used for removing batch.}

\item{only.var.features}{a logical value, whether only return the variable genes.}

\item{add_empty_counts}{a logical value, whether add a empty count matrix. This argument makes it easy to transfer the returned SpatialExperiment object to a Seurat object using \code{\link{as.Seurat}}.}
}
\value{
return a SpatialExperiment object with integrated normalized gene expression matrix and other information from the SRTProject object.
}
\description{
Get the batch-corrected integrated gene expresssions
}
