% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{getSRToutputPrefix}
\alias{getSRToutputPrefix}
\title{Get the Prefix of output information}
\usage{
getSRToutputPrefix()
}
\description{
Get the Prefix of output information
}
\examples{
getSRToutputPrefix()


}
