% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_newinterface.R
\name{Integrate_iSCMEB}
\alias{Integrate_iSCMEB}
\title{Joint clustering and alignment for SRT data by fitting an iSC.MEB model}
\usage{
Integrate_iSCMEB(SRTProj, K = NULL, reduction = "PCA", ...)
}
\arguments{
\item{SRTProj}{an object named "SRTProject". The object SRTProject is created by \link{CreateSRTProject}.}

\item{K}{an optional integer or integer vector, specify the candidates of number of clusters. if K=NULL, it will be set to 4~12.}

\item{reduction}{a string, specify the reduction slot name in SRTProj for fitting iSC.MEB model.}

\item{...}{other parameters pass to \link{model_set_iSCMEB}.}
}
\description{
Joint clustering and alignment for SRT data by fitting an iSC.MEB model
}
\references{
\href{https://doi.org/10.1093/bioadv/vbad019}{ iSC.MEB: an R package for multi-sample spatial clustering analysis of spatial transcriptomics data, Bioinformatics Advances, 2023, vbad019}
}
