% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_newinterface.R
\name{Integrate_PRECAST}
\alias{Integrate_PRECAST}
\title{Joint embedding, clustering and alignment for SRT data by fitting a PRECAST model}
\usage{
Integrate_PRECAST(SRTProj, K = NULL, q = 15, ...)
}
\arguments{
\item{SRTProj}{an object named "SRTProject". The object SRTProject is created by \link{CreateSRTProject}.}

\item{K}{An optional integer or integer vector, specify the candidates of number of clusters. if K=NULL, it will be set to 4~12.}

\item{q}{An optional integer, specify the number of low-dimensional embeddings to extract in PRECAST.}

\item{...}{other parameters pass to \link{model_set_PRECAST}.}
}
\description{
Joint embedding, clustering and alignment for SRT data by fitting a PRECAST model
}
\references{
\href{https://www.nature.com/articles/s41467-023-35947-w}{Wei Liu, Liao, X., Luo, Z. et al, Jin Liu* (2023). Probabilistic embedding, clustering, and alignment for integrating spatial transcriptomics data with PRECAST. Nature Communications, 14, 296}
}
