% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mat2list}
\alias{mat2list}
\title{Transfer a matrix to a list}
\usage{
mat2list(z_int, nvec)
}
\arguments{
\item{z_int}{a vector with number of rows n.}

\item{nvec}{an integer vector, specify the length of each component in the transferred list. \code{n=sum(nvec)}.}
}
\value{
return a list with each componnet matrix.
}
\description{
Transfer a matrix to a list
}
\examples{
mat2list(matrix(1, nrow=3+5, ncol=2), nvec=c(3,5))

}
