% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{matlist2mat}
\alias{matlist2mat}
\title{Transfer matrix list to a matrix}
\usage{
matlist2mat(XList)
}
\arguments{
\item{XList}{a list with each component matrix}
}
\value{
return a rbinded matrix
}
\description{
Transfer matrix list to a matrix
}
\examples{
matlist2mat(list(matrix(1, nrow=2, ncol=2), matrix(2, nrow=4, ncol=2)))


}
