% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_newinterface.R
\name{model_set_DRSC}
\alias{model_set_DRSC}
\title{Set parameters for DR-SC model}
\usage{
model_set_DRSC(
  error_heter = TRUE,
  wpca_int = FALSE,
  int.model = "EEE",
  approxPCA = FALSE,
  beta_grid = seq(0.2, 4, by = 0.2),
  coreNum = 5,
  maxIter_ICM = 6,
  maxIter = 25,
  epsLogLik = 1e-05,
  verbose = FALSE,
  seed = 1
)
}
\arguments{
\item{error_heter}{a logical value, whether use the heterogenous error for DR-SC model, default as TRUE. If error.heter=FALSE, then the homogenuous error is used for probabilistic PCA model in DR-SC.}

\item{wpca_int}{an optional logical value, means whether use the weighted PCA to obtain the initial values of loadings and other paramters, default as FALSE which means the ordinary PCA is used.}

\item{int.model}{an optional string, specify which Gaussian mixture model is used in evaluting the initial values for DR-SC, default as "EEE"; and see Mclust for more models' names.}

\item{approxPCA}{an optional logical value, whether use approximated PCA to speed up the computation for initial values.}

\item{beta_grid}{a numeric vector, specify the smoothness parameter of Random Markov Field. The default is seq(0.2,4,0.2).}

\item{coreNum}{an integer value to decide how many cores are used to run in parallel.}

\item{maxIter_ICM}{the maximum iteration of ICM step, default as 6.}

\item{maxIter}{the maximum iteration of ICM-EM algorithm. The default is 25}

\item{epsLogLik}{an optional positive vlaue, tolerance vlaue of relative variation rate of the observed pseudo log-loglikelihood value, defualt as '1e-5'.}

\item{verbose}{a logical value, whether output the information in iteration.}

\item{seed}{a postive integer, the random seed to be set.}
}
\description{
Prepare parameters setup for DR-SC model fitting.
}
\examples{
model_set_DRSC(approxPCA=TRUE, coreNum=5)

}
