% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_newinterface.R
\name{model_set_iSCMEB}
\alias{model_set_iSCMEB}
\title{Set parameters for iSC.MEB model}
\usage{
model_set_iSCMEB(
  Sigma_equal = FALSE,
  Sigma_diag = TRUE,
  int.model = "EEE",
  beta_grid = seq(0, 5, by = 0.2),
  coreNum = 1,
  maxIter_ICM = 6,
  maxIter = 25,
  epsLogLik = 1e-05,
  verbose = TRUE,
  seed = 1,
  init.start = 1,
  c_penalty = 1
)
}
\arguments{
\item{Sigma_equal}{a logical value, whether set the mixture covariance matrices equal, default as FALSE.}

\item{Sigma_diag}{a logical value, whether set the mixture covariance matrices diagonal, default as TRUE.}

\item{int.model}{an optional string, specify which Gaussian mixture model is used in evaluting the initial values for DR-SC, default as "EEE"; and see Mclust for more models' names.}

\item{beta_grid}{a numeric vector, specify the smoothness parameter of Random Markov Field. The default is seq(0.2,4,0.2).}

\item{coreNum}{an integer value to decide how many cores are used to run  in parallel, default as 5.}

\item{maxIter_ICM}{the maximum iteration of ICM step, default as 6.}

\item{maxIter}{the maximum iteration of ICM-EM algorithm. The default is 25}

\item{epsLogLik}{an optional positive vlaue, tolerance vlaue of relative variation rate of the observed pseudo log-loglikelihood value, defualt as '1e-5'.}

\item{verbose}{a logical value, whether output the information in iteration.}

\item{seed}{a postive integer, the random seed to be set.}

\item{init.start}{a positive integer, the number of times to compute initial values.}

\item{c_penalty}{a positive real, the penalty constant used in modified BIC to determine the number of clusters.}
}
\description{
Prepare parameters setup for iSC.MEB model fitting.
}
\examples{
model_set_iSCMEB(Sigma_equal=TRUE, coreNum=5)

}
