% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_newinterface.R
\name{model_set_Louvain}
\alias{model_set_Louvain}
\title{Set parameters for Louvain model}
\usage{
model_set_Louvain(
  algorithm = 1,
  method = "matrix",
  n.start = 10,
  n.iter = 10,
  seed = 1,
  group.singletons = TRUE
)
}
\arguments{
\item{algorithm}{Algorithm for modularity optimization (1 = original Louvain algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM algorithm; 4 = Leiden algorithm). Leiden requires the leidenalg python.}

\item{method}{Method for running leiden (defaults to matrix which is fast for small datasets). Enable method = "igraph" to avoid casting large data to a dense matrix.}

\item{n.start}{Number of random starts.}

\item{n.iter}{Maximal number of iterations per random start.}

\item{seed}{Seed of the random number generator.}

\item{group.singletons}{Group singletons into nearest cluster. If FALSE, assign all singletons to a "singleton" group.}
}
\description{
Prepare parameters setup for Louvain clustering.
}
\examples{
model_set_Louvain(n.start=5)

}
