% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_newinterface.R
\name{model_set_SCMEB}
\alias{model_set_SCMEB}
\title{Set parameters for SC-MEB model}
\usage{
model_set_SCMEB(
  beta_grid = seq(0.2, 4, by = 0.2),
  coreNum = 1,
  maxIter_ICM = 6,
  maxIter = 30,
  verbose = TRUE,
  seed = 1
)
}
\arguments{
\item{beta_grid}{a numeric vector, specify the smoothness parameter of Random Markov Field. The default is seq(0,4,0.2).}

\item{coreNum}{an integer value to decide how many cores are used to run SC-MEB model in parallel.}

\item{maxIter_ICM}{the maximum iteration of ICM step, default as 6.}

\item{maxIter}{the maximum iteration of ICM-EM algorithm. The default is 30.}

\item{verbose}{a logical value, whether output the information in iteration.}

\item{seed}{a postive integer, the random seed to be set.}
}
\description{
Prepare parameters setup for SC-MEB model fitting.
}
\examples{
model_set_SCMEB(coreNum=5)

}
