% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_newinterface.R
\name{model_set_SpatialAnno}
\alias{model_set_SpatialAnno}
\title{Set parameters for SpatialAnno model}
\usage{
model_set_SpatialAnno(
  error_heter = TRUE,
  wpca_int = TRUE,
  int.model = "SCINA",
  beta_grid = seq(0.1, 2.5, by = 0.2),
  maxIter_ICM = 6,
  maxIter = 30,
  epsLogLik = 1e-05,
  Sigma_diag = FALSE,
  add_Unknown = TRUE,
  verbose = TRUE,
  seed = 1
)
}
\arguments{
\item{error_heter}{a logical value, whether use the heterogenous error for SpatialAnno model, default as TRUE. If error.heter=FALSE, then the homogenuous error is used for probabilistic PCA model in DR-SC.}

\item{wpca_int}{an optional logical value, means whether use the weighted PCA to obtain the initial values of loadings and other paramters, default as FALSE which means the ordinary PCA is used.}

\item{int.model}{an optional string, specify model for computing the initial values for SpatialAnno (one of "SCINA" and "scSorter"), default as "SCINA".}

\item{beta_grid}{a numeric vector, specify the smoothness parameter of Random Markov Field. The default is seq(0.2,4,0.2).}

\item{maxIter_ICM}{the maximum iteration of ICM step, default as 6.}

\item{maxIter}{the maximum iteration of ICM-EM algorithm. The default is 25}

\item{epsLogLik}{an optional positive vlaue, tolerance vlaue of relative variation rate of the observed pseudo log-loglikelihood value, defualt as '1e-5'.}

\item{add_Unknown}{a logical value, whether allow "Unknown" in the identified cell types, default as TRUE.}

\item{verbose}{a logical value, whether output the information in iteration.}

\item{seed}{a postive integer, the random seed to be set.}

\item{approxPCA}{an optional logical value, whether use approximated PCA to speed up the computation for initial values.}
}
\description{
Prepare parameters setup for SpatialAnno model fitting.
}
\examples{
model_set_SpatialAnno(int.model='scSorter')

}
