% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{normalizeSRT}
\alias{normalizeSRT}
\title{Normalize the expression count data}
\usage{
normalizeSRT(
  SRTProj,
  normalization.method = "LogNormalize",
  force = FALSE,
  ...
)
}
\arguments{
\item{SRTProj}{a SRTProject object.}

\item{normalization.method}{a string, specify the method to do normalization.}

\item{force}{a logical value, whether overwrite the data objects in h5file.}

\item{...}{other arguments pass to \code{\link{NormalizeData}}.}
}
\description{
Normalize the expression count data
}
