% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RCTD_run}
\alias{RCTD_run}
\title{Perform spatial deconvolution analysis using RCTD model}
\usage{
RCTD_run(
  sc_data,
  cell_type_columns,
  st_count,
  st_coord,
  output_dir,
  output_name
)
}
\arguments{
\item{sc_data}{a Seurat object, denote the single cell RNA reference panel data.}

\item{cell_type_columns}{a string, the cluster name in the meta data of sc_obj.}

\item{st_count}{a sparse matrix, denote the count matrix of the spatial transcriptomics data to be deconvoluted.}

\item{st_coord}{a matrix with two columns, the spatial coordinates.}

\item{output_dir}{a string, specify the output directory.}

\item{output_name}{a string, the file name used.}
}
\value{
write the results as csv format in the output_dir.
}
\description{
Perform spatial deconvolution analysis using RCTD model
}
