% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RCTD_structure}
\alias{RCTD_structure}
\title{Prepare the reference panel data for the RCTD deconvolution analysis}
\usage{
RCTD_structure(sc_obj, clust_vr)
}
\arguments{
\item{sc_obj}{a Seurat object, denote the single cell RNA reference panel data.}

\item{clust_vr}{a string, the cluster name in the meta data of sc_obj.}
}
\value{
return a list with three components: meta_data, cell_type_dict and dge.
}
\description{
Prepare the reference panel data for the RCTD deconvolution analysis
}
