% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sce2seurat}
\alias{sce2seurat}
\title{SingleCellExperiment object to Seurat object}
\usage{
sce2seurat(sce, verbose = TRUE)
}
\arguments{
\item{sce}{a SingleCellExperiment object, at least including the raw gene count expression matrix.}

\item{verbose}{an optional logical value, whether output the information.}
}
\value{
Return a Seurat object.
}
\description{
Transfer SingleCellExperiment object to a Seurat object for preparation for DR.SC model fitting; see our [DR.SC package website](https://feiyoung.github.io/DR.SC/index.html) for more usage of DR.SC.
}
\examples{
library(SingleCellExperiment)
dir <- system.file(
  file.path("extdata", "10xVisium", "section1"),
  package = "SpatialExperiment")

# read in counts
fnm <- file.path(dir, "raw_feature_bc_matrix")
sce <- DropletUtils::read10xCounts(fnm)
colnames(sce) <- paste0("cell", 1:ncol(spe))


seu <- sce2seurat(sce)
head(seu)

## Fit DR-SC model
library(DR.SC)
library(Seurat)
seu <- NormalizeData(seu)
seu <- FindVariableFeatures(seu)
seu <- DR.SC(seu = seu, K=4, platform = "scRNAseq")
}
