% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{selectVariableFeatures}
\alias{selectVariableFeatures}
\title{Select variable genes}
\usage{
selectVariableFeatures(
  SRTProj,
  nfeatures = 2000,
  type = c("HVGs", "SVGs"),
  method = "vst",
  use_custom_features = NULL,
  exclude_features = NULL
)
}
\arguments{
\item{SRTProj}{a SRTProject object.}

\item{nfeatures}{a positive integer, the number of variable features to be chosen.}

\item{type}{a string, one of "HVGs" and "SVGs", represent the type of variable genes to be chosen.}

\item{method}{a string, the method to use, including 'vst' if \code{type='HVGs'}, "SPARK-X" if \code{type='SVGs'}.}

\item{use_custom_features}{a string vector, a group of custom genes to use. If this argument is not \code{NULL}, it will use this group of custom genes as variable genes.}

\item{exclude_features}{a string vector, a group of genes to be excluded when finding variable genes.}
}
\value{
a revised SRTProject object.
}
\description{
Select variable genes
}
