% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{setSRToutputPrefix}
\alias{setSRToutputPrefix}
\title{Set the Prefix when output informtion}
\usage{
setSRToutputPrefix(SRTprefix = "******")
}
\arguments{
\item{SRTprefix}{a string, default as \code{"******"}.}
}
\description{
Set the Prefix when output informtion
}
\examples{
setSRToutputPrefix()


}
