% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{setSRTverbose}
\alias{setSRTverbose}
\title{Determine whether output the information for all functions in SRTpipeline package.}
\usage{
setSRTverbose(verbose = TRUE)
}
\arguments{
\item{verbose}{a data frame}
}
\description{
Determine whether output the information for all functions in SRTpipeline package.
}
\examples{
setSRTverbose(verbose=FALSE)


}
