% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\docType{class}
\name{SRTProject-class}
\alias{SRTProject-class}
\title{SRTProject Class to store SRT datasets}
\arguments{
\item{projectMetadata}{a SimpleList object to record the project-related info, such as outputPath, projectName}

\item{projectSummary}{a SimpleList object to record the project summary data, such as how many spots for each data batches.}

\item{sampleMetaData}{a DataFrame object to record the info. about each data batch, such as gender, platform and so on.}

\item{cellMetaData}{a DataFrame object to record the meta data of each spot.}

\item{geneMetaData}{a SimpleList object to record the meta data of genes, such as SVGs data.frame, HVGs data.frame.}

\item{spatialCoords}{a DataFrame object to record the spatial coordinates.}

\item{reductions}{a SimpleList to record the inferred dimension reductions (DRs) by dimension reduction methods.}

\item{plotEmbeddings}{a SimpleList to record the embeddings based on DRs, such as tSNE, UMAP.}

\item{clusters}{a factor object to record the inferred clusters by clustering methods.}

\item{models}{a SimpleList to record the model parameters for specified model.}
}
\value{
Returns a UMAP representation.
}
\description{
SRTProject object is a class with a number of slots designed for storing the SRT datasets and the analysis results.
 Key slots to access are listed below.
}
\seealso{
\code{\link{CreateSRTProject}}
}
