% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass.R
\name{topVGs}
\alias{topVGs}
\title{Get the top variables genes}
\usage{
topVGs(SRTProj, ntop = 5, batch = NULL, type = c("HVGs", "SVGs"))
}
\arguments{
\item{SRTProj}{a SRTProject object aftering running \code{\link{selectVariableFeatures}}.}

\item{ntop}{a positive integer, the number of top variable genes.}

\item{batch}{a string, number or NULL, specify the batch name to obtain the variable genes, default as \code{NULL}.}

\item{type}{a string, one of "HVGs" and "SVGs", represent the type of variable genes to be chosen.}
}
\value{
return a string vector, the top variable genes.
}
\description{
Get the top variables genes
}
