% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{TrajectEmbedPlot}
\alias{TrajectEmbedPlot}
\title{Visualize the inferred pseudotime on  embeddings.}
\usage{
TrajectEmbedPlot(
  SRTProj,
  plotEmbeddings = "tSNE",
  no_axis = FALSE,
  base_size = 12,
  border_col = "gray",
  legend.position = "right"
)
}
\arguments{
\item{SRTProj}{a SRTProject object with cellMetaData Traject.commonPseudotime.}

\item{plotEmbeddings}{a string, specify the embeddings.}

\item{no_axis}{a logical value, whether display the axis in plot.}

\item{base_size}{a positive real, the baseline text size in plot.}

\item{border_col}{a string, the color of border in plot.}

\item{legend.position}{a string, the position of legend.}
}
\description{
Visualize the inferred pseudotime on  embeddings.
}
