% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{vec2list}
\alias{vec2list}
\title{Transfer a vector to a list}
\usage{
vec2list(y_int, nvec)
}
\arguments{
\item{y_int}{a vector with length n.}

\item{nvec}{an integer vector, specify the length of each component in the transferred list. \code{n=sum(nvec)}.}
}
\value{
return a list with each component vector
}
\description{
Transfer a vector to a list
}
\examples{
vec2list(c(rep(1,3), rep(2, 5)), nvec=c(3, 5))


}
