% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{write_fig}
\alias{write_fig}
\title{Save a ggplot object to a PNG figure}
\usage{
write_fig(
  plt,
  filename = "myfigure",
  y_reverse = FALSE,
  x_reverse = FALSE,
  width = 7,
  height = 5.5,
  dpi = 200
)
}
\arguments{
\item{plt}{a ggplot object, the object to be saved.}

\item{filename}{a string, the file names of the saved figure, default as `myfigure`.}

\item{y_reverse}{a logical value, whether reverse the graph by y-axis, default as \code{FALSE}.}

\item{x_reverse}{a logical value, whether reverse the graph by x-axis, default as \code{FALSE}.}

\item{width}{a positive real, the width of the saved figure, default as 7.}

\item{height}{a positive real, the height of the saved figure, default as 5.5.}

\item{dpi}{a positive integer, the resolution of the figure, default as 200.}
}
\description{
Save a ggplot object to a PNG figure
}
